migrate to Material stable

+++ src/home/position/positionEditModal/positionEdit.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Output, Input } from '@angular/core';
 import { positionModalHtml } from './positionEdit,html';
 import { PositionService } from '../../../services/position.service';
+import { ToastsManager } from 'ng2-toastr';
 @Component({
   selector: 'position-modal',
   template: positionModalHtml,
@@ -10,7 +11,7 @@ export class PositionModal {
   @Input() position: any;
   @Input() isCreate: boolean;
   @Output() onClose = new EventEmitter();
-  constructor(private positionApi: PositionService) {}
+  constructor(private positionApi: PositionService, private toast: ToastsManager) {}
   ngOnInit() {
     if (!this.position) {
       this.position = {};
@@ -29,10 +30,13 @@ export class PositionModal {
       });
   }
   deletePosition() {
-    this.positionApi.deletePosition(this.position.id)
-      .subscribe(res => {
-        this.onClose.emit(true);
-      });
+    if (confirm('Видалити?')) {
+      this.positionApi.deletePosition(this.position.id)
+        .subscribe(res => {
+          this.toast.success('Успішно видалено');
+          this.onClose.emit(true);
+        });
+    }
   }
   close() {
     this.onClose.emit(false);
